Moving Average and Total
--using AVG( ) And SUM( )--
Moving Average (MA) functions are used  in variety of industries it is the average of the last (n) periods of a column's value
For Example: in sales, the 10 day moving average is the average of the last 10 day's unites sold per day. Its's used to indicate momentum and trends; if a day's unites sold is higher than it's MA , then the next day , more unites are likely to be sold . It is also useful in eliminating seasonality the normal fluctuation of unites sold per day.
And moving total is the SUM of the last (n) periods of a column 's values. It's used to indicate PERFORMANCE  in the recent periods.